:hammer: Use Long type to record clipId (#610)

+++ composeApp/src/commonMain/kotlin/com/clipevery/path/PathProvider.kt
@@ -49,7 +49,7 @@ interface PathProvider {
 
     fun resolve(appInstanceId: String,
                 dateString: String,
-                clipId: Int,
+                clipId: Long,
                 clipFiles: ClipFiles,
                 isPull: Boolean,
                 filesIndexBuilder: FilesIndexBuilder? = null) {
