+++ composeApp/src/desktopMain/kotlin/com/clipevery/utils/IdUtils.kt
@@ -1,18 +1,18 @@
 package com.clipevery.utils
 
 import com.clipevery.dao.clip.ClipDao
-import java.util.concurrent.atomic.AtomicInteger
+import java.util.concurrent.atomic.AtomicLong
 
 class IDGeneratorFactory(private val clipDao: ClipDao) {
 
     fun createIDGenerator(): IDGenerator {
-        val initId = AtomicInteger(clipDao.getMaxClipId())
+        val initId = AtomicLong(clipDao.getMaxClipId())
         return IDGenerator(initId)
     }
 }
 
-class IDGenerator(private val initId: AtomicInteger) {
-    fun nextID(): Int {
+class IDGenerator(private val initId: AtomicLong) {
+    fun nextID(): Long {
         return initId.addAndGet(1)
     }
 }
